requirements.txt
Django==2.2.14
pytz
pyaml
djangorestframework==3.9.4
djangorestframework-xml
djangorestframework-jsonp
icalendar==3.6.2
setuptools==18.5
requests[security]
pysolr==3.2.0
django-haystack
django-bootstrap-form==3.4
psycopg2
raven==6.10.0
git+https://github.com/birniefoxford/oxford-django-shibboleth.git@patch-1#egg=django-shibboleth
python-ldap
django-reversion==2.0.13
textile==2.2.1
django-cors-headers==2.5.0
